+++ src/libs6dns/s6dns-message-internal.h
@@ -3,10 +3,11 @@
 #ifndef S6DNS_MESSAGE_INTERNAL_H
 #define S6DNS_MESSAGE_INTERNAL_H
 
+#include <sys/types.h>
 
  /* Low-level packet parsing */
 
-extern int s6dns_message_get_string_internal (char *, unsigned int, char const *, unsigned int, unsigned int *) ;
-extern unsigned int s6dns_message_get_domain_internal (char *, unsigned int, char const *, unsigned int, unsigned int *) ;
+extern int s6dns_message_get_string_internal (char *, size_t, char const *, unsigned int, unsigned int *) ;
+extern size_t s6dns_message_get_domain_internal (char *, size_t, char const *, unsigned int, unsigned int *) ;
 
 #endif
