+++ src/libs6dns/s6dns_domain_fromstring_qualify_encode.c
@@ -2,7 +2,7 @@
 
 #include <s6-dns/s6dns-domain.h>
 
-unsigned int s6dns_domain_fromstring_qualify_encode (s6dns_domain_t *list, char const *name, unsigned int len, char const *rules, unsigned int rulesnum)
+unsigned int s6dns_domain_fromstring_qualify_encode (s6dns_domain_t *list, char const *name, size_t len, char const *rules, unsigned int rulesnum)
 {
   s6dns_domain_t d ;
   if (!s6dns_domain_fromstring(&d, name, len)) return 0 ;
