+++ src/libs6dns/s6dns_domain_tostring.c
@@ -1,10 +1,11 @@
 /* ISC license. */
 
+#include <sys/types.h>
 #include <errno.h>
 #include <skalibs/bytestr.h>
 #include <s6-dns/s6dns-domain.h>
 
-unsigned int s6dns_domain_tostring (char *s, unsigned int max, s6dns_domain_t const *d)
+unsigned int s6dns_domain_tostring (char *s, size_t max, s6dns_domain_t const *d)
 {
   if ((unsigned int)d->len + 1 > max) return (errno = ENAMETOOLONG, 0) ;
   if (!d->len || (d->s[0] != '.')) return (errno = EINVAL, 0) ;
